+++ src/main/java/com/unstoppabledomains/resolution/contracts/HTTPUtil.java
@@ -2,16 +2,6 @@ package com.unstoppabledomains.resolution.contracts;
 
 import com.google.gson.JsonArray;
 import com.google.gson.JsonObject;
-import com.google.gson.JsonParser;
-import com.unstoppabledomains.config.client.Client;
-
-import java.io.BufferedReader;
-import java.io.IOException;
-import java.io.InputStreamReader;
-import java.io.OutputStream;
-import java.net.HttpURLConnection;
-import java.net.URL;
-import java.nio.charset.StandardCharsets;
 
 public class HTTPUtil {
     private HTTPUtil() {
@@ -25,38 +15,5 @@ public class HTTPUtil {
         body.add("params", params);
         return body;
     }
-
-    public static JsonObject post(String url, JsonObject body) throws IOException {
-        HttpURLConnection con = HTTPUtil.createAndConfigureCon(url);
-        try (OutputStream os = con.getOutputStream()) {
-            byte[] input = body.toString().getBytes(StandardCharsets.UTF_8);
-            os.write(input, 0, input.length);
-        }
-
-        try (BufferedReader br = new BufferedReader(new InputStreamReader(con.getInputStream(), StandardCharsets.UTF_8))) {
-            StringBuilder response = new StringBuilder();
-            String responseLine;
-            while ((responseLine = br.readLine()) != null) {
-                response.append(responseLine.trim());
-            }
-            return (JsonObject) JsonParser.parseString(response.toString());
-        }
-    }
-
-    private static HttpURLConnection createAndConfigureCon(String url) throws IOException {
-        URL posturl = new URL(url);
-        HttpURLConnection con = (HttpURLConnection) posturl.openConnection();
-        con.setRequestMethod("POST");
-        con.setRequestProperty("Content-Type", "application/json");
-        con.setRequestProperty("Accept", "application/json");
-        con.addRequestProperty("User-Agent", getUserAgent());
-        con.setDoOutput(true);
-        return con;
-    }
-
-    private static String getUserAgent() {
-        String agent = "UnstoppableDomains/resolution-java";
-        String version = Client.getVersion();
-        return version.isEmpty() ? agent : agent + "/" + version;
-    }
+    
 }