+++ src/main/java/com/unstoppabledomains/resolution/naming/service/BaseNamingService.java
@@ -1,13 +1,16 @@
 package com.unstoppabledomains.resolution.naming.service;
 
 import com.unstoppabledomains.config.network.model.Network;
+import com.unstoppabledomains.resolution.contracts.interfaces.IProvider;
 
 public abstract class BaseNamingService implements NamingService {
     protected Network chainId;
     protected String blockchainProviderUrl;
+    protected IProvider provider;
 
-    protected BaseNamingService(NSConfig nsConfig) {
+    protected BaseNamingService(NSConfig nsConfig, IProvider provider) {
         this.chainId = nsConfig.getChainId();
         this.blockchainProviderUrl = nsConfig.getBlockchainProviderUrl();
+        this.provider = provider;
     }
 }