+++ src/test/java/com/unstoppabledomains/TestUtils.java
@@ -21,4 +21,21 @@ public class TestUtils {
             assertEquals(code, e.getCode());
         }
     }
+
+    public static void checkError(Callable<String> f, NSExceptionCode code, Throwable cause) throws Exception {
+        try {
+            f.call();
+            System.out.println("Didn't throw an " + code + " exception");
+            fail();
+        } catch (NamingServiceException e) {
+            if (code != e.getCode()) {
+                throw e;
+            }
+            if (cause != e.getCause()) {
+                throw e;
+            }
+            assertEquals(code, e.getCode());
+            assertEquals(cause, e.getCause());
+        }
+    }
 }